Update script.js/index.html
Updated the script.js to have a variable for the submit button/must add the id, added a taskTitleInput.value for a tasks array, and added an on click call the handleAddTask function. Updated the index.html file to have a new button layout for the Add Task button, and added a modal with a form inside of it, and changed the classes for the labels as well as what they're for, will change them so they match up to javascript.