Update index.html/script.js
Updated index.html to remove a redundant code on line 25, changed id tags on line 30 then on line 34 then on line 42 then on line 46. Updated the script.js to have variables linked to the id tags for my form. Cleaned up the rest of the code, and changed the generate task id function, and moved the date picker function.