Update script.js/index.html
Updated index.html to have the button on line 52 target the body class on line 37 so I can select that button in javascript, and generate a unique task id when it is clicked. Updated the script.js file to clean the code a bit, and started on the function generateTaskId.