Update script.js
updated the create task card function to take a task object as an argument to create a string for a task card. The task includes the title due date description, and a delete button. The function now calculates the remaining days until the task's due date and assigns a color class based on this number. The rendertasklist function now empties the columns then loops through the task list array and creates a task card for each task. However the function is currently incomplete and doesnt append the card to the page or make them draggable so I will be adding these in with a TA before my tutor session tomorrow.