Update script.js/index.html
Updated the index.html to have new comments, and updated the script.js to comment out the const declarations at on lines 4-8, line 41 is now seperated to lines 47-49, removed line 67 and now the function for the date picker is inside the document.ready function on line 72. However I will work on adding the rest with my tutor so I can get some notes.