Update script.js
Updated script.js on line 2 added an array for the tasks. on line 5 uncommented the submitButton selector. On line 18 calculate the number of days remaining until the task is due. On line 30 added a class of draggable, then on line 35 added quotation to the task.id to target the task id. Lines 48-67 added the create task card render to make it so you can drag the card when it is rendered. Line 73-92 prevent default so the card can render, made a tasks object to push to the taskList, and stored it to the localStorage taskList then rendered the tasklist. lines 96-104 added code to work on commented it out it was interfering with other code. lines 112-115 Added a droppable function to for the drag and drop feature.