Update script.js
Updated script.js on lines 95-103 so that there is a task Id that is targeted on line 96, then on 97 the task id is filtered and parsed to then set the object into local storage. Then the taskList is rendered with the taskId to be then selected on line 100 then the button delete can be clicked to delete the targeted object(s). on lines 131-134 I uncommented the jquery method to drop taskList object.