Update script.js
Updated script.js to make the task cards persist on refresh, However I have not commented it yet. Updated the colors to change accordingly to the progress state they are in, However I have not commented it yet, so I'm going to comment that, and also changed the delete button to not delete closest instead delete by getelementclassname however its not functioning yet so I will fix that soon.