Update README.md/script.js
Updated the README.md to correct a spelling error, change dates on my sources, changed the description, changed the usage, added screenshots, and live link now functions. Updated script.js to make the handleDeleteTask function operate correctly now the tasks are deleted and do not persist when reloading the page.